Update qase-go repo (#280)
sdg

Update schema.go

Update qase.go

+++ actions/qase/defaults.go
@@ -2,7 +2,7 @@ package qase
 
 const (
 	AutomationSuiteID       = int32(554)
-	AutomationTestNameID    = 15
+	AutomationTestNameID    = int64(15)
 	QaseTokenEnvVar         = "QASE_AUTOMATION_TOKEN"
 	RancherManagerProjectID = "RM"
 	RecurringRunID          = 1
